requirements.txt

awsebcli==3.13.0
Django==1.11.13
flickrapi==2.4.0
Pillow==5.1.0
requests==2.9.1
requests-file==1.4.2
requests-ftp==0.3.1
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
urllib3==1.22

